+++ docs/plugins/feed/config/README.md
@@ -145,7 +145,7 @@ Atom syntax output filename, relative to dest folder.
 ## atomXslTemplate
 
 - Type: `string`
-- Default: Content of `vuepress-plugin-feed2/templates/atom.xsl`
+- Default: Content of `@vuepress/plugin-feed/templates/atom.xsl`
 
 Atom xsl template file content.
 
@@ -173,7 +173,7 @@ RSS syntax output filename, relative to dest folder.
 ## rssXslTemplate
 
 - Type: `string`
-- Default: Content of `vuepress-plugin-feed2/templates/rss.xsl`
+- Default: Content of `@vuepress/plugin-feed/templates/rss.xsl`
 
 RSS xsl template file content.
