+++ plugins/plugin-feed/src/node/feed/store.ts
@@ -1,11 +1,11 @@
 import type { App } from 'vuepress/core'
-import type { FeedLinks, ResolvedFeedOptions } from '../options.js'
-import { getFeedChannelOption, getFeedLinks } from '../options.js'
 import type {
   FeedCategory,
   FeedChannelOption,
   FeedContributor,
-} from '../typings/index.js'
+} from '../../typings/index.js'
+import type { FeedLinks, ResolvedFeedOptions } from '../options.js'
+import { getFeedChannelOption, getFeedLinks } from '../options.js'
 import type { FeedItem } from './item.js'
 
 export class FeedStore {
