+++ plugins/plugin-feed/src/node/generator/atom/index.ts
@@ -1,7 +1,7 @@
 import { isArray } from '@vuepress/helper/node'
 import { js2xml } from 'xml-js'
+import type { FeedAuthor, FeedCategory } from '../../../typings/index.js'
 import type { FeedStore } from '../../feed/store.js'
-import type { FeedAuthor, FeedCategory } from '../../typings/index.js'
 import { encodeXML, FEED_GENERATOR } from '../../utils/index.js'
 import type {
   AtomAuthor,
