+++ plugins/plugin-feed/src/node/generator/json/index.ts
@@ -1,6 +1,6 @@
 import { isArray } from '@vuepress/helper/node'
+import type { FeedAuthor } from '../../../typings/index.js'
 import type { FeedStore } from '../../feed/store.js'
-import type { FeedAuthor } from '../../typings/index.js'
 import type { JSONAuthor, JSONContent, JSONItem } from './typings.js'
 
 const getJSONAuthor = (author: FeedAuthor): JSONAuthor => ({
