+++ plugins/plugin-feed/src/node/getFeed.ts
@@ -2,6 +2,7 @@ import { entries, fromEntries } from '@vuepress/helper/node'
 import type { GitData } from '@vuepress/plugin-git'
 import type { App, Page } from 'vuepress/core'
 import { colors } from 'vuepress/utils'
+import type { FeedConfig, FeedPluginFrontmatter } from '../typings/index.js'
 import { FeedItem } from './feed/item.js'
 import { FeedStore } from './feed/store.js'
 import { getAtomFeed } from './generator/atom/index.js'
@@ -9,7 +10,6 @@ import { getJSONFeed } from './generator/json/index.js'
 import { getRssFeed } from './generator/rss/index.js'
 import type { ResolvedFeedOptionsMap } from './options.js'
 import { getFilename } from './options.js'
-import type { FeedConfig, FeedPluginFrontmatter } from './typings/index.js'
 import { logger } from './utils/index.js'
 
 export const getFeedFiles = (
