+++ plugins/plugin-feed/src/node/index.ts
@@ -1,2 +1,2 @@
-export * from './plugin.js'
-export * from './typings/index.js'
+export * from './feedPlugin.js'
+export * from '../typings/index.js'
